moved DabHandler and FlossDanceHandler to better package
