+++ src/main/java/cc/hyperium/handlers/HyperiumHandlers.java
@@ -24,6 +24,8 @@ import cc.hyperium.event.InvokeEvent;
 import cc.hyperium.event.TickEvent;
 import cc.hyperium.gui.ScoreboardRenderer;
 import cc.hyperium.handlers.handlers.*;
+import cc.hyperium.handlers.handlers.animation.DabHandler;
+import cc.hyperium.handlers.handlers.animation.FlossDanceHandler;
 import cc.hyperium.handlers.handlers.chat.*;
 import cc.hyperium.handlers.handlers.keybinds.KeyBindHandler;
 import cc.hyperium.handlers.handlers.privatemessages.PrivateMessageHandler;
